fixed a bug where changing a sample packs icon with the change image button would crash the game if there wasn't already an icon there

+++ src/client/java/com/provinceofmusic/screen/SamplePackEditor.java
@@ -93,7 +93,9 @@ public class SamplePackEditor extends LightweightGuiDescription {
                     File file = dialog.getFiles()[0];
 
                     try {
-                        Files.delete(Path.of(SamplePack.getFile(thisPack.name).toPath() + "\\icon.png"));
+                        if(new File(SamplePack.getFile(thisPack.name).toPath() + "\\icon.png").exists()){
+                            Files.delete(Path.of(SamplePack.getFile(thisPack.name).toPath() + "\\icon.png"));
+                        }
                         Files.copy(file.toPath(), Path.of(SamplePack.getFile(thisPack.name).toPath() + "\\icon.png"));
                     } catch (IOException e) {
                         throw new RuntimeException(e);
